Module google_texttospeech1::api

Structs

AudioConfig
Description of audio data to be synthesized.

ListVoicesResponse
The message returned to the client by the ListVoices method.

SynthesisInput
Contains text input to be synthesized. Either text or ssml must be supplied. Supplying both or neither returns google.rpc.Code.INVALID_ARGUMENT. The input size is limited to 5000 characters.

SynthesizeSpeechRequest
The top-level message sent by the client for the SynthesizeSpeech method.

SynthesizeSpeechResponse
The message returned to the client by the SynthesizeSpeech method.

TextMethods
A builder providing access to all methods supported on text resources. It is not used directly, but through the Texttospeech hub.

TextSynthesizeCall
Synthesizes speech synchronously: receive results after all text input has been processed.

Texttospeech
Central instance to access all Texttospeech related resource activities

Voice
Description of a voice supported by the TTS service.

VoiceListCall
Returns a list of Voice supported for synthesis.

VoiceMethods
A builder providing access to all methods supported on voice resources. It is not used directly, but through the Texttospeech hub.

VoiceSelectionParams
Description of which voice to use for a synthesis request.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.